refactor(config): move cinemaserver to own file
